Cleaned up documentation and code.
